-Added dataextractor test changes.

+++ dtests/src/test/java/io/snappydata/hydra/cdcConnector/SnappyCDCTest.java
@@ -319,7 +319,7 @@ public class SnappyCDCTest extends SnappyTest {
         String cntQry = "SELECT COUNT(*) FROM " + tableName;
         ResultSet rs3 = con.createStatement().executeQuery(cntQry);
         while (rs3.next()) {
-          count = rs3.getInt(1);
+          count = rs3.getLong(1);
           Log.getLogWriter().info("SP: The tableName = " + tableName + " count = " + count);
         }
         if(!tableCntMap.containsKey(tableName)) //To avoid overwriting the existing key(table)
